+++ src/main/java/com/gregtechceu/gtceu/api/mui/widgets/SchemaWidget.java
@@ -1,47 +1,52 @@
 package com.gregtechceu.gtceu.api.mui.widgets;
 
-import com.gregtechceu.gtceu.api.mui.base.drawable.IDrawable;
 import com.gregtechceu.gtceu.api.mui.base.drawable.IKey;
 import com.gregtechceu.gtceu.api.mui.base.widget.Interactable;
-import com.gregtechceu.gtceu.api.mui.drawable.SchemaRenderer;
+import com.gregtechceu.gtceu.api.mui.theme.WidgetTheme;
 import com.gregtechceu.gtceu.api.mui.widget.Widget;
+import com.gregtechceu.gtceu.client.mui.screen.viewport.ModularGuiContext;
 import com.gregtechceu.gtceu.common.mui.GTGuiTextures;
+import com.gregtechceu.gtceu.utils.GTMath;
 import com.gregtechceu.gtceu.utils.VectorUtil;
+import com.gregtechceu.gtceu.utils.fakelevel.BaseSchemaRenderer;
 import com.gregtechceu.gtceu.utils.fakelevel.ISchema;
 
 import net.minecraft.network.chat.Component;
 import net.minecraft.util.Mth;
 
 import org.jetbrains.annotations.NotNull;
-import org.jetbrains.annotations.Nullable;
 import org.joml.Vector3f;
 
-import static net.minecraft.util.Mth.PI;
 import static net.minecraft.util.Mth.TWO_PI;
 
 public class SchemaWidget extends Widget<SchemaWidget> implements Interactable {
 
-    private final SchemaRenderer schema;
+    private final BaseSchemaRenderer schema;
     private boolean enableRotation = true;
     private boolean enableTranslation = true;
     private boolean enableScaling = true;
     private float lastMouseX;
     private float lastMouseY;
-    private double scale = 10;
-    private float pitch = (float) (Math.PI / 4f);
-    private float yaw = (float) (Math.PI / 4f);
+    private float scale = 10f;
+    private float pitch = GTMath.PI_QUART;
+    private float yaw = 0;
     private final Vector3f offset = new Vector3f();
 
     public SchemaWidget(ISchema schema) {
-        this(new SchemaRenderer(schema));
+        this(new BaseSchemaRenderer(schema));
     }
 
-    public SchemaWidget(SchemaRenderer schema) {
+    public SchemaWidget(BaseSchemaRenderer schema) {
         this.schema = schema;
         schema.cameraFunc((camera, $schema) -> {
             Vector3f focus = VectorUtil.vec3fAdd(this.offset, null, $schema.getFocus());
-            camera.setLookAt(focus, scale, yaw, pitch);
+            camera.setLookAtAndAngle(focus, scale, yaw, pitch);
         });
+    }
+
+    @Override
+    public void draw(ModularGuiContext context, WidgetTheme widgetTheme) {
+        this.schema.drawAtZero(context, getArea(), widgetTheme);
     }
 
     @Override
@@ -67,22 +72,23 @@ public class SchemaWidget extends Widget<SchemaWidget> implements Interactable {
         float dx = (float) mX - lastMouseX;
         float dy = (float) mY - lastMouseY;
         if (button == 0 && this.enableRotation) {
-            float moveScale = 0.025f;
+            float moveScale = 0.03f;
             yaw = (yaw + dx * moveScale + TWO_PI) % TWO_PI;
             pitch = Mth.clamp(pitch + dy * moveScale, -TWO_PI / 4 + 0.001f, TWO_PI / 4 - 0.001f);
         } else if (button == 2 && this.enableTranslation) {
             // the idea is to construct a vector which points upwards from the camera pov (y-axis on screen)
             // this vector determines the amount of z offset from mouse movement in y
             float y = (float) Math.cos(pitch);
-            float moveScale = 0.06f;
+            float moveScale = 0.09f;
             // with this the offset can be moved by dy
             offset.add(0, dy * y * moveScale, 0);
             // to respect dx we need a new vector which is perpendicular on the previous vector (x-axis on screen)
             // y = 0 => mouse movement in x does not move y
-            float phi = (yaw + PI / 2) % TWO_PI;
-            float x = (float) Math.cos(phi);
-            float z = (float) Math.sin(phi);
-            offset.add(dx * x * moveScale, 0, dx * z * moveScale);
+            Vector3f look = this.schema.camera().getLookVec().normalize(); // direction camera is looking
+            Vector3f right = look.cross(GTMath.UNIT_Y).normalize(); // right relative to screen
+            Vector3f up = right.cross(look);
+            this.offset.add(right.mul(-dx * moveScale)).add(up.mul(dy * moveScale));
+
         }
         this.lastMouseX = (float) mX;
         this.lastMouseY = (float) mY;
@@ -90,16 +96,6 @@ public class SchemaWidget extends Widget<SchemaWidget> implements Interactable {
 
     public SchemaWidget scale(double scale) {
         this.scale += scale;
-        return this;
-    }
-
-    public SchemaWidget pitch(float pitch) {
-        this.pitch += pitch;
-        return this;
-    }
-
-    public SchemaWidget yaw(float yaw) {
-        this.yaw += yaw;
         return this;
     }
 
@@ -131,11 +127,6 @@ public class SchemaWidget extends Widget<SchemaWidget> implements Interactable {
 
     public SchemaWidget enableAllInteraction(boolean enable) {
         return enableInteraction(enable, enable, enable);
-    }
-
-    @Override
-    public @Nullable IDrawable getOverlay() {
-        return schema;
     }
 
     public static class LayerButton extends ButtonWidget<LayerButton> {
