+++ src/main/java/com/gregtechceu/gtceu/common/machine/muimachine/TestMuiMachine.java
@@ -37,6 +37,9 @@ import com.gregtechceu.gtceu.client.mui.screen.RichTooltip;
 import com.gregtechceu.gtceu.client.mui.screen.UISettings;
 import com.gregtechceu.gtceu.client.mui.screen.viewport.ModularGuiContext;
 import com.gregtechceu.gtceu.common.mui.GTGuiTextures;
+import com.gregtechceu.gtceu.utils.fakelevel.ArraySchema;
+import com.gregtechceu.gtceu.utils.fakelevel.BaseSchemaRenderer;
+import com.gregtechceu.gtceu.utils.fakelevel.BlockHighlight;
 
 import net.minecraft.network.chat.Component;
 import net.minecraft.world.entity.EntityType;
@@ -538,14 +541,12 @@ public class TestMuiMachine extends MetaMachine implements IMuiMachine {
         page.sizeRel(1f);
         page.child(IKey.str("Schema").asWidget());
 
-        /*
-         * if (getLevel().isClientSide()) {
-         * page.child(new SchemaWidget(new SchemaRenderer(ArraySchema.of(data.getPlayer(), 5))
-         * .highlightRenderer(new BlockHighlight(Color.withAlpha(Color.GREEN.brighter(1), 0.9f), 1 / 32f)))
-         * .pos(20, 20)
-         * .size(100, 100));
-         * }
-         */
+        if (getLevel().isClientSide()) {
+            page.child(new SchemaWidget(new BaseSchemaRenderer(ArraySchema.of(data.getPlayer(), 5))
+                    .highlightRenderer(new BlockHighlight(Color.withAlpha(Color.GREEN.brighter(1), 0.9f), 1 / 32f)))
+                    .pos(20, 20)
+                    .size(100, 100));
+        }
 
         return page;
     }
