+++ src/main/java/com/gregtechceu/gtceu/utils/fakelevel/BaseSchemaRenderer.java
@@ -2,35 +2,95 @@ package com.gregtechceu.gtceu.utils.fakelevel;
 
 import com.gregtechceu.gtceu.api.mui.base.drawable.IDrawable;
 import com.gregtechceu.gtceu.api.mui.drawable.Icon;
+import com.gregtechceu.gtceu.api.mui.theme.WidgetTheme;
+import com.gregtechceu.gtceu.api.mui.utils.Color;
+import com.gregtechceu.gtceu.api.mui.widget.sizer.Area;
 import com.gregtechceu.gtceu.api.mui.widgets.SchemaWidget;
-import lombok.Getter;
-import net.minecraft.world.level.BlockGetter;
-import net.minecraft.world.level.Level;
-import net.minecraft.world.phys.HitResult;
-import org.joml.Vector3f;
+import com.gregtechceu.gtceu.client.mui.screen.viewport.GuiContext;
+import com.gregtechceu.gtceu.utils.GTMatrixUtils;
 
+import net.minecraft.client.Minecraft;
+import net.minecraft.client.renderer.GameRenderer;
+import net.minecraft.client.renderer.MultiBufferSource;
+import net.minecraft.client.renderer.RenderType;
+import net.minecraft.client.renderer.block.BlockRenderDispatcher;
+import net.minecraft.client.renderer.blockentity.BlockEntityRenderDispatcher;
+import net.minecraft.core.BlockPos;
+import net.minecraft.util.RandomSource;
+import net.minecraft.world.inventory.InventoryMenu;
+import net.minecraft.world.level.ClipContext;
+import net.minecraft.world.level.LevelReader;
+import net.minecraft.world.level.block.RenderShape;
+import net.minecraft.world.level.block.entity.BlockEntity;
+import net.minecraft.world.level.block.state.BlockState;
+import net.minecraft.world.phys.HitResult;
+import net.minecraft.world.phys.Vec3;
+import net.minecraftforge.client.model.data.ModelData;
+
+import com.mojang.blaze3d.pipeline.RenderTarget;
+import com.mojang.blaze3d.pipeline.TextureTarget;
+import com.mojang.blaze3d.platform.Lighting;
+import com.mojang.blaze3d.systems.RenderSystem;
+import com.mojang.blaze3d.vertex.*;
+import lombok.Getter;
+import lombok.Setter;
+import lombok.experimental.Accessors;
+import lombok.experimental.Tolerate;
+import org.joml.Matrix4f;
+import org.joml.Vector3f;
+import org.lwjgl.opengl.GL11;
+
+import java.util.Objects;
+import java.util.function.BiConsumer;
+import java.util.function.BooleanSupplier;
+import java.util.function.Consumer;
+import java.util.function.DoubleSupplier;
+
+@Accessors(fluent = true)
 public class BaseSchemaRenderer implements IDrawable {
 
-    private static final Framebuffer FBO = new Framebuffer(1080, 1080, true);
+    private static final RenderTarget FBO = new TextureTarget(1080, 1080, true, Minecraft.ON_OSX);
 
-    @Getter
-    private final ISchema schema;
-    private final BlockGetter renderWorld;
-    private final Framebuffer framebuffer;
+    protected final ISchema schema;
+    private final LevelReader renderLevel;
+    private final RenderTarget renderTarget;
     @Getter
     private final Camera camera = new Camera();
-    private HitResult lastRayTrace = null;
+    @Setter
+    private boolean cameraSetup = false;
+    @Setter
+    protected DoubleSupplier scale;
+    @Setter
+    protected BooleanSupplier disableBER;
+    @Setter
+    protected Consumer<IRayTracer> onRayTrace;
+    @Setter
+    protected Runnable afterRender;
+    @Setter
+    protected BiConsumer<Camera, ISchema> cameraFunc;
+    private int clearColor = 0;
+    @Setter
+    protected boolean isometric = false;
 
-    public BaseSchemaRenderer(ISchema schema, Framebuffer framebuffer) {
+    public BaseSchemaRenderer(ISchema schema, RenderTarget renderTarget) {
         this.schema = schema;
-        this.framebuffer = framebuffer;
-        this.renderWorld = new renderWorld.;
+        this.renderTarget = renderTarget;
+        this.renderLevel = new RenderLevel(schema);
     }
 
     public BaseSchemaRenderer(ISchema schema) {
         this(schema, FBO);
     }
 
+    @Tolerate
+    public BaseSchemaRenderer scale(double scale) {
+        return scale(() -> scale);
+    }
+
+    @Tolerate
+    public BaseSchemaRenderer disableBER(boolean disable) {
+        return disableBER(() -> disable);
+    }
 
     @Override
     public SchemaWidget asWidget() {
@@ -42,248 +102,214 @@ public class BaseSchemaRenderer implements IDrawable {
         return IDrawable.super.asIcon().size(50);
     }
 
+    @Override
+    public void draw(GuiContext context, int x, int y, int width, int height, WidgetTheme widgetTheme) {
+        render(context, x, y, width, height, context.getMouseX(), context.getMouseY());
+    }
 
-    protected void render(int x, int y, int width, int height, int mouseX, int mouseY) {
-        onSetupCamera();
-        int lastFbo = bindFBO();
-        setupCamera(this.framebuffer.framebufferWidth, this.framebuffer.framebufferHeight);
-        renderWorld();
-        if (doRayTrace()) {
-            HitResult result = null;
-            if (Area.isInside(x, y, width, height, mouseX, mouseY)) {
-                result = rayTrace(mouseX, mouseY, width, height);
-            }
-            if (result == null) {
-                if (this.lastRayTrace != null) {
-                    onRayTraceFailed();
-                }
-            } else {
-                onSuccessfulRayTrace(result);
-            }
-            this.lastRayTrace = result;
+    public void render(GuiContext context, int x, int y, int width, int height, int mouseX, int mouseY) {
+        if (this.cameraFunc != null) {
+            this.cameraFunc.accept(this.camera, this.schema);
         }
-        onRendered();
-        Platform.setupDrawTex();
+        if (Objects.nonNull(scale)) {
+            Vector3f cameraPos = camera.pos();
+            Vector3f looking = camera.lookAt();
+            cameraPos.sub(looking);
+            if (cameraPos.length() != 0.0f) cameraPos.normalize();
+            cameraPos.mul((float) scale.getAsDouble());
+            cameraPos.add(looking);
+        }
+
+        this.renderTarget.setClearColor(0.0F, 0.0F, 0.0F, 0.0F);
+        this.renderTarget.clear(Minecraft.ON_OSX);
+        this.renderTarget.bindWrite(true);
+
+        context.getGraphics().pose().pushPose();
+        setupCamera(this.renderTarget.viewWidth, this.renderTarget.viewHeight);
+        renderWorld(context);
+        if (this.onRayTrace != null && Area.isInside(x, y, width, height, mouseX, mouseY)) {
+            this.onRayTrace.accept(new IRayTracer() {
+
+                @Override
+                public HitResult rayTrace(int screenX, int screenY) {
+                    return BaseSchemaRenderer.this.rayTrace(GTMatrixUtils.projectScreenToWorld(screenX, screenY));
+                }
+
+                @Override
+                public HitResult rayTraceMousePos() {
+                    return rayTrace(mouseX, mouseY);
+                }
+            });
+        }
         resetCamera();
-        unbindFBO(lastFbo);
+        context.getGraphics().pose().popPose();
+        this.renderTarget.unbindWrite();
+        Minecraft.getInstance().getMainRenderTarget().bindWrite(true);
 
         // bind FBO as texture
-        GlStateManager.enableTexture2D();
-        GlStateManager.disableLighting();
-        lastFbo = GL11.glGetInteger(GL11.GL_TEXTURE_2D);
-        GlStateManager.bindTexture(this.framebuffer.framebufferTexture);
-        GlStateManager.color(1, 1, 1, 1);
+        RenderSystem.setShader(GameRenderer::getPositionTexShader);
+        RenderSystem.setShaderTexture(0, this.renderTarget.getColorTextureId());
+        RenderSystem.setShaderColor(1, 1, 1, 1);
 
         // render rect with FBO texture
-        Platform.startDrawing(Platform.DrawMode.QUADS, Platform.VertexFormat.POS_TEX, bufferBuilder -> {
-            bufferBuilder.pos(x + width, y + height, 0).tex(1, 0).endVertex();
-            bufferBuilder.pos(x + width, y, 0).tex(1, 1).endVertex();
-            bufferBuilder.pos(x, y, 0).tex(0, 1).endVertex();
-            bufferBuilder.pos(x, y + height, 0).tex(0, 0).endVertex();
-        });
-        GlStateManager.bindTexture(lastFbo);
+        Tesselator tesselator = Tesselator.getInstance();
+        BufferBuilder bufferbuilder = tesselator.getBuilder();
+        bufferbuilder.begin(VertexFormat.Mode.QUADS, DefaultVertexFormat.POSITION_TEX);
+
+        bufferbuilder.vertex(x + width, y + height, 0).uv(1, 0).endVertex();
+        bufferbuilder.vertex(x + width, y, 0).uv(1, 1).endVertex();
+        bufferbuilder.vertex(x, y, 0).uv(0, 1).endVertex();
+        bufferbuilder.vertex(x, y + height, 0).uv(0, 0).endVertex();
+        tesselator.end();
     }
 
-    protected HitResult rayTrace(int mouseX, int mouseY, int width, int height) {
-        final float halfPI = (float) (Math.PI / 2);
-        Vector3f cameraPos = camera.getPos();
-        float yaw = camera.getYaw();
-        float pitch = camera.getPitch();
+    private void renderWorld(GuiContext context) {
+        PoseStack poseStack = context.getGraphics().pose();
+        MultiBufferSource.BufferSource bufferSource = context.getGraphics().bufferSource();
+        RandomSource random = RandomSource.create();
 
-        Vector3f mouseXShift = new Vector3f(1, 0, 0)
-                .rotatePitch(pitch)
-                .rotateYaw(-yaw + halfPI)
-                .scale(mouseX - width / 2f)
-                .scale(1 / 32f);
-        Vector3f mouseYShift = new Vector3f(0, -1, 0)
-                .rotatePitch(pitch)
-                .rotateYaw(-yaw + halfPI)
-                .scale(mouseY - height / 2f)
-                .scale(1 / 32f);
-        Vector3f mousePos = Vector3f.add(cameraPos, mouseXShift, mouseYShift, null);
-        Vector3f focus = camera.getLookAt();
-        float perspectiveCompensation = isIsometric() ? 1 : cameraPos.distanceTo(focus) / 3 * width / 100;
-        Vector3f underMousePos = focus.add(mouseXShift.scale(perspectiveCompensation), null).add(mouseYShift.scale(perspectiveCompensation));
-        Vector3f look = Vector3f.sub(underMousePos, mousePos, null).scale(10);
-        Vector3f.add(mousePos, look, underMousePos);
-        return schema.getWorld().rayTraceBlocks(mousePos.toVec3d(), underMousePos.toVec3d(), true);
-    }
+        Minecraft mc = Minecraft.getInstance();
+        Vec3 cameraPos = mc.gameRenderer.getMainCamera().getPosition();
+        RenderSystem.enableCull();
+        Lighting.setupForFlatItems();
+        mc.gameRenderer.lightTexture().turnOffLightLayer();
+        RenderSystem.setShaderTexture(0, InventoryMenu.BLOCK_ATLAS);
 
-    private void renderWorld() {
-        Minecraft mc = Minecraft.getMinecraft();
-        Platform.setupDrawTex();
-        GlStateManager.enableCull();
-        GlStateManager.enableRescaleNormal();
-        RenderHelper.disableStandardItemLighting();
-        mc.entityRenderer.disableLightmap();
-        mc.renderEngine.bindTexture(TextureMap.LOCATION_BLOCKS_TEXTURE);
-        BlockRenderLayer oldRenderLayer = MinecraftForgeClient.getRenderLayer();
-        GlStateManager.disableLighting();
-        Platform.setupDrawGradient(); // needed for ambient occlusion
+        for (final RenderType type : RenderType.chunkBufferLayers()) {
+            type.setupRenderState();
+            BufferBuilder buffer = Tesselator.getInstance().getBuilder();
+            buffer.begin(VertexFormat.Mode.QUADS, DefaultVertexFormat.BLOCK);
+            BlockRenderDispatcher blockRenderer = mc.getBlockRenderer();
+            this.schema.forEach(pair -> {
+                BlockPos pos = pair.getKey();
+                BlockState state = pair.getValue().getBlockState();
+                if (state.getRenderShape() == RenderShape.INVISIBLE) return;
 
-        try { // render block in each layer
-            for (BlockRenderLayer layer : BlockRenderLayer.values()) {
-                ForgeHooksClient.setRenderLayer(layer);
-                int pass = layer == BlockRenderLayer.TRANSLUCENT ? 1 : 0;
-                setDefaultPassRenderState(pass);
-                BufferBuilder buffer = Tessellator.getInstance().getBuffer();
-                buffer.begin(GL11.GL_QUADS, DefaultVertexFormats.BLOCK);
-                BlockRendererDispatcher blockrendererdispatcher = mc.getBlockRendererDispatcher();
-                this.schema.forEach(pair -> {
-                    BlockPos pos = pair.getKey();
-                    IBlockState state = pair.getValue().getBlockState();
-                    if (!state.getBlock().isAir(state, this.renderWorld, pos) && state.getBlock().canRenderInLayer(state, layer)) {
-                        blockrendererdispatcher.renderBlock(state, pos, this.renderWorld, buffer);
-                    }
-                });
-                Tessellator.getInstance().draw();
-                Tessellator.getInstance().getBuffer().setTranslation(0, 0, 0);
-            }
-        } finally {
-            ForgeHooksClient.setRenderLayer(oldRenderLayer);
+                var be = pair.getValue().getBlockEntity();
+                ModelData modelData = ModelData.EMPTY;
+                if (be != null) {
+                    modelData = be.getModelData();
+                }
+                poseStack.pushPose();
+                poseStack.translate(-cameraPos.x(), -cameraPos.y(), -cameraPos.z());
+                blockRenderer.renderBatched(state, pos, this.renderLevel,
+                        poseStack, buffer, true,
+                        random, modelData, type);
+                poseStack.popPose();
+
+            });
+            Tesselator.getInstance().end();
+            type.clearRenderState();
         }
 
-        RenderHelper.enableStandardItemLighting();
-        GlStateManager.enableLighting();
+        Lighting.setupFor3DItems();
 
-        // render TESR
-        if (isTesrEnabled()) {
-            for (int pass = 0; pass < 2; pass++) {
-                ForgeHooksClient.setRenderPass(pass);
-                int finalPass = pass;
-                GlStateManager.color(1, 1, 1, 1);
-                setDefaultPassRenderState(pass);
-                this.schema.forEach(pair -> {
-                    BlockPos pos = pair.getKey();
-                    TileEntity tile = pair.getValue().getTileEntity();
-                    if (tile != null && tile.shouldRenderInPass(finalPass)) {
-                        TileEntityRendererDispatcher.instance.render(tile, pos.getX(), pos.getY(), pos.getZ(), 0);
-                    }
-                });
-            }
+        // render BERs
+        if (disableBER == null || !disableBER.getAsBoolean()) {
+            RenderSystem.setShaderColor(1, 1, 1, 1);
+            BlockEntityRenderDispatcher blockEntityRenderer = mc.getBlockEntityRenderDispatcher();
+            float partialTick = mc.getPartialTick();
+            this.schema.forEach(pair -> {
+                BlockPos pos = pair.getKey();
+                if (pair.getValue().getBlockState().getRenderShape() == RenderShape.MODEL) return;
+                BlockEntity be = pair.getValue().getBlockEntity();
+                if (be == null) return;
+                poseStack.pushPose();
+                poseStack.translate(pos.getX() - cameraPos.x(), pos.getY() - cameraPos.y(), pos.getZ() - cameraPos.z());
+                blockEntityRenderer.render(be, partialTick, poseStack, bufferSource);
+                poseStack.popPose();
+            });
         }
-        Platform.endDrawGradient();
-        ForgeHooksClient.setRenderPass(-1);
-        GlStateManager.enableDepth();
-        GlStateManager.disableBlend();
-        GlStateManager.depthMask(true);
-    }
-
-    private static void setDefaultPassRenderState(int pass) {
-        GlStateManager.color(1, 1, 1, 1);
-        if (pass == 0) { // SOLID
-            GlStateManager.enableDepth();
-            GlStateManager.disableBlend();
-            GlStateManager.depthMask(true);
-        } else { // TRANSLUCENT
-            GlStateManager.enableBlend();
-            GlStateManager.blendFunc(GL11.GL_SRC_ALPHA, GL11.GL_ONE_MINUS_SRC_ALPHA);
-            GlStateManager.depthMask(false);
+        RenderSystem.enableDepthTest();
+        RenderSystem.disableBlend();
+        RenderSystem.depthMask(true);
+        if (this.afterRender != null) {
+            this.afterRender.run();
         }
     }
 
-    protected final void setupCamera(int width, int height) {
-        //GlStateManager.pushAttrib();
-
-        Minecraft.getMinecraft().entityRenderer.disableLightmap();
-        GlStateManager.disableLighting();
-        GlStateManager.enableDepth();
-        GlStateManager.enableBlend();
+    protected void setupCamera(int width, int height) {
+        Minecraft.getInstance().gameRenderer.lightTexture().turnOffLightLayer();
+        RenderSystem.enableDepthTest();
+        RenderSystem.enableBlend();
 
         // setup viewport and clear GL buffers
-        GlStateManager.viewport(0, 0, width, height);
-        Color.setGlColor(getClearColor());
-        GlStateManager.clear(GL11.GL_COLOR_BUFFER_BIT | GL11.GL_DEPTH_BUFFER_BIT);
+        RenderSystem.viewport(0, 0, width, height);
+        Color.setGlColor(clearColor);
+        RenderSystem.clear(GL11.GL_COLOR_BUFFER_BIT | GL11.GL_DEPTH_BUFFER_BIT, Minecraft.ON_OSX);
+        RenderSystem.backupProjectionMatrix();
 
-        // setup projection matrix to perspective
-        GlStateManager.matrixMode(GL11.GL_PROJECTION);
-        GlStateManager.pushMatrix();
-        GlStateManager.loadIdentity();
-
-        float near = isIsometric() ? 1f : 0.1f;
+        float near = this.isometric ? 1f : 0.1f;
         float far = 10000.0f;
         float fovY = 60.0f; // Field of view in the Y direction
         float aspect = (float) width / height; // width and height are the dimensions of your window
-        float top = near * (float) Math.tan(Math.toRadians(fovY) / 2.0);
-        float bottom = -top;
-        float left = aspect * bottom;
-        float right = aspect * top;
-        if (isIsometric()) {
-            GL11.glOrtho(left, right, bottom, top, near, far);
+        Matrix4f projection = new Matrix4f();
+        if (this.isometric) {
+            float top = near * (float) Math.tan(Math.toRadians(fovY) / 2.0);
+            float bottom = -top;
+            float left = aspect * bottom;
+            float right = aspect * top;
+            projection.setOrtho(left, right, bottom, top, near, far);
+            RenderSystem.setProjectionMatrix(projection, VertexSorting.ORTHOGRAPHIC_Z);
         } else {
-            GL11.glFrustum(left, right, bottom, top, near, far);
+            projection.setPerspective(fovY, aspect, near, far);
+            RenderSystem.setProjectionMatrix(projection, VertexSorting.byDistance(camera.lookAt()));
         }
 
         // setup modelview matrix
-        GlStateManager.matrixMode(GL11.GL_MODELVIEW);
-        GlStateManager.pushMatrix();
-        GlStateManager.loadIdentity();
-        if (isIsometric()) {
-            GlStateManager.scale(0.1, 0.1, 0.1);
+        PoseStack modelViewStack = RenderSystem.getModelViewStack();
+        modelViewStack.pushPose();
+        modelViewStack.setIdentity();
+        if (this.isometric) {
+            modelViewStack.scale(0.1f, 0.1f, 0.1f);
         }
-        var c = this.camera.getPos();
-        var lookAt = this.camera.getLookAt();
-        GLU.gluLookAt(c.x, c.y, c.z, lookAt.x, lookAt.y, lookAt.z, 0, 1, 0);
+        var cameraPos = this.camera.pos();
+        var lookAt = this.camera.lookAt();
+        modelViewStack.mulPoseMatrix(GTMatrixUtils.lookAt(cameraPos, lookAt));
+        RenderSystem.applyModelViewMatrix();
+        this.cameraSetup = true;
     }
 
-    protected final void resetCamera() {
+    protected void resetCamera() {
+        this.cameraSetup = false;
         // reset viewport
-        Minecraft minecraft = Minecraft.getMinecraft();
-        GlStateManager.viewport(0, 0, minecraft.displayWidth, minecraft.displayHeight);
+        Minecraft minecraft = Minecraft.getInstance();
+        RenderSystem.viewport(0, 0, minecraft.getWindow().getScreenWidth(), minecraft.getWindow().getScreenHeight());
+
         // reset projection matrix
-        GlStateManager.matrixMode(GL11.GL_PROJECTION);
-        GlStateManager.popMatrix();
+        RenderSystem.restoreProjectionMatrix();
+
         // reset modelview matrix
-        GlStateManager.matrixMode(GL11.GL_MODELVIEW);
-        GlStateManager.popMatrix();
-        GlStateManager.disableBlend();
-        GlStateManager.disableDepth();
-        // reset attributes
-        // GlStateManager.popAttrib();
+        RenderSystem.getModelViewStack().popPose();
+        RenderSystem.applyModelViewMatrix();
+
+        RenderSystem.disableBlend();
+        RenderSystem.disableDepthTest();
     }
 
-    private int bindFBO() {
-        int lastID = GL11.glGetInteger(EXTFramebufferObject.GL_FRAMEBUFFER_BINDING_EXT);
-        this.framebuffer.setFramebufferColor(0.0F, 0.0F, 0.0F, 0.0F);
-        this.framebuffer.framebufferClear();
-        this.framebuffer.bindFramebuffer(true);
-        GlStateManager.pushMatrix();
-        return lastID;
+    private HitResult rayTrace(Vector3f hitPos) {
+        Vec3 startPos = new Vec3(this.camera.pos().x, this.camera.pos().y, this.camera.pos().z);
+        hitPos.mul(2); // Double view range to ensure pos can be seen.
+        Vec3 endPos = new Vec3((hitPos.x - startPos.x), (hitPos.y - startPos.y), (hitPos.z - startPos.z));
+
+        ClipContext context = new ClipContext(startPos, endPos,
+                ClipContext.Block.COLLIDER, ClipContext.Fluid.NONE, null);
+        return this.schema.getLevel().clip(context);
     }
 
-    private void unbindFBO(int lastID) {
-        GlStateManager.popMatrix();
-        this.framebuffer.unbindFramebufferTexture();
-        OpenGlHelper.glBindFramebuffer(OpenGlHelper.GL_FRAMEBUFFER, lastID);
+    public boolean isCameraSetup() {
+        return cameraSetup;
     }
 
-    @ApiStatus.OverrideOnly
-    protected void onSetupCamera() {}
+    public interface IRayTracer {
 
-    @ApiStatus.OverrideOnly
-    protected void onRendered() {}
+        HitResult rayTrace(int screenX, int screenY);
 
-    @ApiStatus.OverrideOnly
-    protected void onSuccessfulRayTrace(@NotNull RayTraceResult result) {}
-
-    @ApiStatus.OverrideOnly
-    protected void onRayTraceFailed() {}
-
-    public boolean doRayTrace() {
-        return false;
+        HitResult rayTraceMousePos();
     }
 
-    public int getClearColor() {
-        return Color.withAlpha(0, 0.5f);
-    }
+    public interface ICamera {
 
-    public boolean isIsometric() {
-        return false;
-    }
-
-    public boolean isTesrEnabled() {
-        return true;
+        void setupCamera(Vector3f cameraPos, Vector3f lookAt);
     }
 }
-
-
